tempo/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
gunicorn
python-dotenv
jsonify
Flask-JSON
redis
flask-cors
marshmallow-sqlalchemy
keybert
celery
tensorflow
torch==2.0.*